+++ src/Summary.cpp
@@ -4,8 +4,8 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2016-2017 XMRig       <support@xmrig.com>
- *
+ * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -23,6 +23,7 @@
 
 
 #include <inttypes.h>
+#include <stdio.h>
 #include <uv.h>
 
 
@@ -33,18 +34,18 @@
 #endif
 
 
+#include "common/log/Log.h"
+#include "common/net/Pool.h"
+#include "core/Config.h"
+#include "core/Controller.h"
 #include "Cpu.h"
-#include "log/Log.h"
-#include "net/Url.h"
-#include "Options.h"
 #include "Summary.h"
 #include "version.h"
-#include "workers/OclThread.h"
 
 
-static void print_versions()
+static void print_versions(xmrig::Config *config)
 {
-    char buf[16];
+    char buf[16] = { 0 };
 
 #   if defined(__clang__)
     snprintf(buf, 16, " clang/%d.%d.%d", __clang_major__, __clang_minor__, __clang_patchlevel__);
@@ -52,10 +53,7 @@ static void print_versions()
     snprintf(buf, 16, " gcc/%d.%d.%d", __GNUC__, __GNUC_MINOR__, __GNUC_PATCHLEVEL__);
 #   elif defined(_MSC_VER)
     snprintf(buf, 16, " MSVC/%d", MSVC_VERSION);
-#   else
-    buf[0] = '\0';
 #   endif
-
 
 #   if CL_VERSION_2_0
     const char *ocl = "2.0";
@@ -69,89 +67,100 @@ static void print_versions()
     const char *ocl = "0.0";
 #   endif
 
-    Log::i()->text(Options::i()->colors() ? "\x1B[01;32m * \x1B[01;37mVERSIONS:     \x1B[01;36mXMRig/%s\x1B[01;37m libuv/%s OpenCL/%s%s" : " * VERSIONS:     XMRig/%s libuv/%s OpenCL/%s%s",
-                   APP_VERSION, uv_version_string(), ocl, buf);
+    Log::i()->text(config->isColors() ? GREEN_BOLD(" * ") WHITE_BOLD("%-13s") CYAN_BOLD("%s/%s") WHITE_BOLD(" libuv/%s OpenCL/%s%s")
+                                      : " * %-13s%s/%s libuv/%s OpenCL/%s%s",
+                   "VERSIONS", APP_NAME, APP_VERSION, uv_version_string(), ocl, buf);
 }
 
 
-static void print_cpu()
+static void print_cpu(xmrig::Config *config)
 {
-    if (Options::i()->colors()) {
-        Log::i()->text("\x1B[01;32m * \x1B[01;37mCPU:          %s %sx64 %sAES-NI",
+    if (config->isColors()) {
+        Log::i()->text(GREEN_BOLD(" * ") WHITE_BOLD("%-13s") WHITE_BOLD("%s %sx64 %sAES"),
+                       "CPU",
                        Cpu::brand(),
-                       Cpu::isX64() ? "\x1B[01;32m" : "\x1B[01;31m-",
-                       Cpu::hasAES() ? "\x1B[01;32m" : "\x1B[01;31m-");
+                       Cpu::isX64() ? "\x1B[1;32m" : "\x1B[1;31m-",
+                       Cpu::hasAES() ? "\x1B[1;32m" : "\x1B[1;31m-");
     }
     else {
-        Log::i()->text(" * CPU:          %s (%d) %sx64 %sAES-NI", Cpu::brand(), Cpu::sockets(), Cpu::isX64() ? "" : "-", Cpu::hasAES() ? "" : "-");
+        Log::i()->text(" * %-13s%s %sx64 %sAES", "CPU", Cpu::brand(), Cpu::isX64() ? "" : "-", Cpu::hasAES() ? "" : "-");
     }
 }
 
 
-static void print_algo()
+static void print_algo(xmrig::Config *config)
 {
-    Log::i()->text(Options::i()->colors() ? "\x1B[01;32m * \x1B[01;37mALGO:         %s, %sdonate=%d%%" : " * ALGO:         %s, %sdonate=%d%%",
-                   Options::i()->algoName(),
-                   Options::i()->colors() && Options::i()->donateLevel() == 0 ? "\x1B[01;31m" : "",
-                   Options::i()->donateLevel()
+    Log::i()->text(config->isColors() ? GREEN_BOLD(" * ") WHITE_BOLD("%-13s%s, %sdonate=%d%%")
+                                      : " * %-13s%s, %sdonate=%d%%",
+                   "ALGO",
+                   config->algorithm().name(),
+                   config->isColors() && config->donateLevel() == 0 ? "\x1B[1;31m" : "",
+                   config->donateLevel()
     );
 }
 
 
-static void print_pools()
+static void print_pools(xmrig::Config *config)
 {
-    const std::vector<Url*> &pools = Options::i()->pools();
+    const std::vector<Pool> &pools = config->pools();
 
     for (size_t i = 0; i < pools.size(); ++i) {
-        Log::i()->text(Options::i()->colors() ? "\x1B[01;32m * \x1B[01;37mPOOL #%d:      \x1B[01;36m%s:%d" : " * POOL #%d:      %s:%d",
+        Log::i()->text(config->isColors() ? GREEN_BOLD(" * ") WHITE_BOLD("POOL #%-7zu") CYAN_BOLD("%s") " variant " WHITE_BOLD("%s")
+                                          : " * POOL #%-7d%s variant %s",
                        i + 1,
-                       pools[i]->host(),
-                       pools[i]->port());
+                       pools[i].url(),
+                       pools[i].algorithm().variantName()
+                       );
     }
 
 #   ifdef APP_DEBUG
-    for (size_t i = 0; i < pools.size(); ++i) {
-        Log::i()->text("%s:%d, user: %s, pass: %s, ka: %d, nicehash: %d", pools[i]->host(), pools[i]->port(), pools[i]->user(), pools[i]->password(), pools[i]->isKeepAlive(), pools[i]->isNicehash());
+    for (const Pool &pool : pools) {
+        pool.print();
     }
 #   endif
 }
 
 
 #ifndef XMRIG_NO_API
-static void print_api()
+static void print_api(xmrig::Config *config)
 {
-    if (Options::i()->apiPort() == 0) {
+    const int port = config->apiPort();
+    if (port == 0) {
         return;
     }
 
-    Log::i()->text(Options::i()->colors() ? "\x1B[01;32m * \x1B[01;37mAPI PORT:     \x1B[01;36m%d" : " * API PORT:     %d", Options::i()->apiPort());
+    Log::i()->text(config->isColors() ? GREEN_BOLD(" * ") WHITE_BOLD("%-13s") CYAN("%s:") CYAN_BOLD("%d")
+                                      : " * %-13s%s:%d",
+                   "API BIND", config->isApiIPv6() ? "[::]" : "0.0.0.0", port);
 }
 #endif
 
 
-static void print_commands()
+static void print_commands(xmrig::Config *config)
 {
-    if (Options::i()->colors()) {
-        Log::i()->text("\x1B[01;32m * \x1B[01;37mCOMMANDS:     \x1B[01;35mh\x1B[01;37mashrate, \x1B[01;35mp\x1B[01;37mause, \x1B[01;35mr\x1B[01;37mesume");
+    if (config->isColors()) {
+        Log::i()->text(GREEN_BOLD(" * ") WHITE_BOLD("COMMANDS     ") MAGENTA_BOLD("h") WHITE_BOLD("ashrate, ")
+                                                                     MAGENTA_BOLD("p") WHITE_BOLD("ause, ")
+                                                                     MAGENTA_BOLD("r") WHITE_BOLD("esume"));
     }
     else {
-        Log::i()->text(" * COMMANDS:     'h' hashrate, 'p' pause, 'r' resume");
+        Log::i()->text(" * COMMANDS     'h' hashrate, 'p' pause, 'r' resume");
     }
 }
 
 
-void Summary::print()
+void Summary::print(xmrig::Controller *controller)
 {
-    print_versions();
-    print_cpu();
-    print_algo();
-    print_pools();
+    print_versions(controller->config());
+    print_cpu(controller->config());
+    print_algo(controller->config());
+    print_pools(controller->config());
 
 #   ifndef XMRIG_NO_API
-    print_api();
+    print_api(controller->config());
 #   endif
 
-    print_commands();
+    print_commands(controller->config());
 }