+++ src/amd/OclCLI.h
@@ -4,8 +4,8 @@
  * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2016-2017 XMRig       <support@xmrig.com>
- *
+ * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -28,7 +28,16 @@
 #include <vector>
 
 
+#include "common/xmrig.h"
+
+
 class OclThread;
+
+
+namespace xmrig {
+    class Config;
+    class IThread;
+}
 
 
 class OclCLI
@@ -36,8 +45,8 @@ class OclCLI
 public:
     OclCLI();
 
-    bool setup(std::vector<OclThread*> &threads);
-    void autoConf(std::vector<OclThread*> &threads, int *platformIndex);
+    bool setup(std::vector<xmrig::IThread *> &threads);
+    void autoConf(std::vector<xmrig::IThread *> &threads, int *platformIndex, xmrig::Config *config);
     void parseLaunch(const char *arg);
 
     inline void parseAffinity(const char *arg) { parse(m_affinity, arg); }